// api/mock-server/src/generated/mock-admin/contract/types.gen.ts
export type FolderRecord = DomainsFoldersFolder & {
    id?: string;
    deletedAt?: string;
};

export type DeckRecord = DomainsDecksDeck & {
    id?: string;
    deletedAt?: string;
};

export type NoteDetailRecord = DomainsNotesNoteDetail & {
    id?: string;
    deletedAt?: string;
};
// ...
export type VisualIconName = string;

export type Id = string;

export type DateTime = string;

export type ActiveWorkspace = DomainsWorkspacesActiveWorkspace;

export type DomainsWorkspacesActiveWorkspace = {
    workspaceId: Id;
};

export type Folder = DomainsFoldersFolder;

export type DomainsFoldersFolder = {
    description: string;
    id: Id;
    name: string;
    parentId: Id;
    updatedAt: DateTime;
    workspaceId: Id;
};

export type Deck = DomainsDecksDeck;

export type DomainsDecksDeck = {
    description: string;
    dueToday: number;
    icon: VisualIconName;
    id: Id;
    parentId: Id;
    progress: number;
    title: string;
    totalNotes: number;
    updatedAt: DateTime;
    workspaceId: Id;
};

export type NoteDetail = DomainsNotesNoteDetail;

export type DomainsNotesNoteDetail = ({
    kind: 'basic';
} & BasicNote) | ({
    kind: 'cloze';
} & ClozeNote);

export type BasicNote = {
    deckId: Id;
    dueAt: DateTime;
    editor: BasicNoteEditor;
    id: Id;
    kind: 'basic';
    progress: number;
    reviewedAt: DateTime;
    status: NoteStatus;
    title: string;
    updatedAt: DateTime;
};

export type BasicNoteEditor = {
    back: string;
    front: string;
};

export type NoteStatus = 'in-progress' | 'mastered';

export type ClozeNote = {
    deckId: Id;
    cards: Array<ClozeNoteCard>;
    dueAt: DateTime;
    editor: ClozeNoteEditor;
    id: Id;
    kind: 'cloze';
    progress: number;
    reviewedAt: DateTime;
    status: NoteStatus;
    title: string;
    updatedAt: DateTime;
};